展開プログラムのリネーム処理見直し

+++ Source/Applications/sbin/Extractor/MainWorkerViewModel.cs
@@ -56,6 +56,8 @@ namespace ContentTypeTextNet.MnMn.SystemApplications.Extractor
         TextWriter Writer { get; set; }
         string LogFilePath { get; set; }
 
+        bool IsRenamed { get; set; }
+
         public CollectionModel<LogItemViewModel> LogItems { get; } = new CollectionModel<LogItemViewModel>();
 
         public bool CanInput
@@ -272,11 +274,14 @@ namespace ContentTypeTextNet.MnMn.SystemApplications.Extractor
             var myDir = Path.GetDirectoryName(myPath);
 
             var renamePath = Path.ChangeExtension(myPath, "expand-old");
-            if(File.Exists(renamePath)) {
-                File.Delete(renamePath);
+            if(!IsRenamed) {
+                if(File.Exists(renamePath)) {
+                    File.Delete(renamePath);
+                }
+                AddInformationLog($"Rename: {myPath} => {renamePath}");
+                File.Move(myPath, renamePath);
+                IsRenamed = true;
             }
-            AddInformationLog($"Rename: {myPath} => {renamePath}");
-            File.Move(myPath, renamePath);
 
             // 置き換え開始
             using(var archive = new ZipArchive(new FileStream(ArchiveFilePath, FileMode.Open, FileAccess.Read, FileShare.None), ZipArchiveMode.Read)) {
